Factor out interrupt priority header
